+++ spec/integration/resource/api300/c7000/logical_switch/update_spec.rb
@@ -9,9 +9,9 @@ RSpec.describe klass, integration: true, type: UPDATE do
     @item.retrieve!
   end
 
-  describe '#refresh' do
+  describe '#refresh_data_state!' do
     it 'refresh logical switch' do
-      @item.refresh
+      @item.refresh_data_state!
     end
   end
 end